requirements.txt
# Core dependencies
grobid-client-python>=0.0.7
lxml>=4.9.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ML/AI dependencies
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
accelerate>=0.24.0

# Vector database
qdrant-client>=1.7.0

# NoSQL database
cassandra-driver>=3.28.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6

# API framework (optional, for service)
fastapi>=0.104.0
uvicorn>=0.24.0
flask>=3.0.0
flask-cors>=4.0.0


